+++ kinto/plugins/history/listener.py
@@ -78,7 +78,7 @@ def on_resource_changed(event):
         perms = {k: list(v) for k, v in perms_by_object_id[uri].items()}
         eventattrs = dict(**payload)
         eventattrs.pop('bucket_id', None)
-        eventattrs.setdefault('%s_id' % resource_name, obj_id)
+        eventattrs['%s_id' % resource_name] = obj_id
         eventattrs['uri'] = uri
         attrs = dict(date=datetime.now().isoformat(),
                      target={'data': target, 'permissions': perms},